Refactory contacts and roles

+++ resources/views/roles/index.blade.php
@@ -25,24 +25,32 @@
 
 <table class="table table-bordered">
   <tr>
-     <th>No</th>
+     <th>ID</th>
      <th>Name</th>
      <th width="280px">Action</th>
   </tr>
     @foreach ($roles as $key => $role)
     <tr>
-        <td>{{ ++$i }}</td>
+        <td>{{ $role->id }}</td>
         <td>{{ $role->name }}</td>
         <td>
             <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
             @can('role-edit')
                 <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
             @endcan
-            @can('role-delete')
-                {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                {!! Form::close() !!}
-            @endcan
+            @if($role->id  != 1)
+                @can('role-delete')
+                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                    {!! Form::close() !!}
+                @endcan
+            @else
+                @can('role-delete')
+                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger', 'disabled']) !!}
+                    {!! Form::close() !!}
+                @endcan
+            @endif
         </td>
     </tr>
     @endforeach
